+++ Extension/ExtensionInterface.php
@@ -30,20 +30,6 @@ interface ExtensionInterface
     public function load(array $configs, ContainerBuilder $container): void;
 
     /**
-     * Returns the namespace to be used for this extension (XML namespace).
-     *
-     * @deprecated since Symfony 7.4, to be removed in Symfony 8.0 together with XML support.
-     */
-    public function getNamespace(): string;
-
-    /**
-     * Returns the base path for the XSD files.
-     *
-     * @deprecated since Symfony 7.4, to be removed in Symfony 8.0 together with XML support.
-     */
-    public function getXsdValidationBasePath(): string|false;
-
-    /**
      * Returns the recommended alias to use in XML.
      *
      * This alias is also the mandatory prefix to use when using YAML.